Final design implementation of datepicker and scss issues fixes

+++ projects/ui/src/lib/oui/datepicker/native-date-formats.ts
@@ -1,11 +1,3 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
 import { OuiDateFormats } from './date-formats';
 
 export const OUI_NATIVE_DATE_FORMATS: OuiDateFormats = {
@@ -13,7 +5,7 @@ export const OUI_NATIVE_DATE_FORMATS: OuiDateFormats = {
     dateInput: null
   },
   display: {
-    dateInput: { year: 'numeric', month: 'numeric', day: 'numeric' },
+    dateInput: { year: 'numeric', month: 'long', day: 'numeric' },
     monthYearLabel: { year: 'numeric', month: 'short' },
     dateA11yLabel: { year: 'numeric', month: 'long', day: 'numeric' },
     monthYearA11yLabel: { year: 'numeric', month: 'long' }
